Implementing JS/TS/logic-related PR feedback.
