--- packages/webapp/src/inductions/hooks/queries.ts
@@ -1,8 +1,0 @@
-import { useQuery } from "react-query";
-import { getInductionWithEndorsements } from "inductions/api";
-
-export const useGetInductionWithEndorsements = (inductionId: string) =>
-    useQuery(["induction", inductionId], getInductionWithEndorsements, {
-        // gets rid of bugs cause by refetching right after anchor/scatter signing
-        refetchOnWindowFocus: false,
-    });
